requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
pydicom
python-multipart
